# go/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ticker": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert tickers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "ticker",
                "responses": {}
            }
        },
        "/ticker/daycandle": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert daycandles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dayCandle"
                ],
                "summary": "day candles",
                "responses": {}
            }
        },
        "/ticker/daycandle/kr": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert daycandles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dayCandle"
                ],
                "summary": "day candles",
                "responses": {}
            }
        },
        "/ticker/daycandle/kr/index": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert index daycandles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dayCandle"
                ],
                "summary": "index day candles",
                "responses": {}
            }
        },
        "/ticker/kr": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert kr tickers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "krTicker",
                "responses": {}
            }
        },
        "/ticker/kr/trend": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "insert ticker trend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticker"
                ],
                "summary": "kr ticker trend",
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}